feat(#937): implement sys_readdir pagination with kernel-owned PaginatedResult
- Add core/pagination.py with PaginatedResult + paginate_iter (kernel primitives)
- sys_readdir returns PaginatedResult when limit is provided, plain list otherwise
- Move PaginatedResult/paginate_iter from lib/ to core/ (syscall return type belongs in kernel)
- lib/pagination.py retains cursor encode/decode for brick-layer use
- Un-skip test_list_pagination.py — all 16 tests pass
- Fix pre-existing test bug: recursive=False doesn't list subdirectories
